+++ src/main/types.ts
@@ -390,6 +390,7 @@ type ObsAudioConfig = {
   obsForceMono: boolean;
   speakerVolume: number;
   micVolume: number;
+  processVolume: number;
   pushToTalk: boolean;
   pushToTalkKey: number;
   pushToTalkMouseButton: number;
@@ -557,6 +558,20 @@ export declare type TObsFormData = (
   | IObsListInput<TObsValue>
 )[];
 
+type ObsSourceCallbackInfo = {
+  name: string;
+  width: number;
+  height: number;
+  flags: number;
+};
+
+type ObsVolmeterCallbackInfo = {
+  sourceName: string;
+  magnitude: number[];
+  peak: number[];
+  inputPeak: number[];
+};
+
 export {
   RecStatus,
   SaveStatus,
@@ -610,4 +625,6 @@ export {
   CreateMultiPartUploadResponseBody,
   CompleteMultiPartUploadRequestBody,
   StorageFilter,
+  ObsSourceCallbackInfo,
+  ObsVolmeterCallbackInfo,
 };